MMUMerge patch from Oliver Ford.
Adds the vars MMUMergeStart, MMUMergeCount, MMUMergeTargetStart and
MMUMergeTargetSize. The specified entries of the level 1 table are
watched during WIRQ and the number of times they change recorded.

On WIRQ completion, each entry's last non-null change is shown
including any level 2 entries if the L2 table still exists in
memory. If specified, only entries with physical addresses in the
target range are shown.