Add INITRD_OFFSET variable (more info in commit)
This new variable works in the same way as KERNEL_OFFSET. It can be set to contain a delta
offset to change the default value. By default:
 * Tags are placed at  RAM START + KERNEL_OFFSET + 0x100
 * Kernel is placed at RAM START + KERNEL_OFFSET + 0x8000
 * Initrd is placed at RAM START + KERNEL_OFFSET + 0x8000 + 0x500000 + INITRD_OFFSET

So, 5MB is reserved for the kernel by default. If your kernel is larger than 5MB, use the
INITRD_OFFSET variable to move the initrd a bit further so that it doesn't overlap the kernel
code.