[ARMv7] Fixed flush cache for ARMv7
The ARMv7 flush cache function was flawed. It would only flush the cache
partially. This caused some last-millisecond writes in HaRET not to be
persisted to RAM and to not be available when the MMU was turned off.

The ARMv7 flush cache function uses the CLZ instruction. In order to
make use of this instruction, which was added in the ARMv5 instruction
set, the source had to be compiled with --march=armv5. In order to
retain backward compatibility, the asmstuff.S file has been split
into an armv4 and an armv5 file. The asmstuff-armv5.S may only
contain code that is called when it is established the current
architecture supports it.

Makefile has been updated to compile any files with "-armv5." in the
filename with the --march=armv5 compile flags.